+++ Firestore/Example/Tests/Integration/API/FIRWriteBatchTests.mm
@@ -249,7 +249,7 @@
   FIRDocumentReference *doc = [self documentRef];
   FSTEventAccumulator *accumulator = [FSTEventAccumulator accumulatorForTest:self];
   [doc
-      addSnapshotListenerWithOptions:[[FIRDocumentListenOptions options] includeMetadataChanges:YES]
+      addSnapshotListenerWithIncludeMetadataChanges:YES
                             listener:accumulator.valueEventHandler];
   FIRDocumentSnapshot *initialSnap = [accumulator awaitEventWithName:@"initial event"];
   XCTAssertFalse(initialSnap.exists);